Working Version

+++ spring-boot-jdbc-example/src/main/java/com/efreelearn/dao/SubjectInfoRepository.java
@@ -1,11 +1,14 @@
 package com.efreelearn.dao;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.efreelearn.model.SubjectInfo;
-@Repository
-public interface SubjectInfoRepository extends JpaRepository<SubjectInfo,Integer>{
+
+import org.springframework.data.jpa.repository.JpaRepository;
+@Repository   
+public interface SubjectInfoRepository extends JpaRepository<SubjectInfo, Long>   
+{
+	
 }
